pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
virt-maker = "virtmaker.cli:cli"

[project]
name = "virt-maker"
version = "1.0.45"
authors = [
  { name="Josiah Kerley", email="josiahkerley@gmail.com" },
]
description = "A virtual machine building tool built on the libguestfs tools."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Natural Language :: English",
    "Development Status :: 3 - Alpha"
]
dependencies = [
    "colorama==0.4.6",
    "Jinja2==3.1.2",
    "PyYAML==6.0.1",
    "psutil==5.9.5",
    "vncdotool==1.2.0",
    "pytesseract==0.3.10",
    "pillow==10.3.0",
    "jsonschema==4.21.1",
    "jsonschema-specifications==2023.12.1",
    "fuzzywuzzy==0.18.0"
]

[project.urls]
"Homepage" = "https://github.com/JosiahKerley/virt-maker"
"Bug Tracker" = "https://github.com/JosiahKerley/virt-maker/issues"

